refactor(middleware): organize utility functions
Creates `utils` package
